move LAI to canopy directory
