+++ docs/src/tutorials/calibration/perfect_model_site_level_calibration.jl
@@ -110,8 +110,11 @@ forcing = FluxnetSimulations.prescribed_forcing_fluxnet(
 );
 
 # Get Leaf Area Index (LAI) data from MODIS satellite observations.
-LAI =
-    ClimaLand.prescribed_lai_modis(domain.space.surface, start_date, stop_date);
+LAI = ClimaLand.Canopy.prescribed_lai_modis(
+    domain.space.surface,
+    start_date,
+    stop_date,
+);
 
 # ## Model Setup
 #
