+++ docs/src/tutorials/integrated/fluxnet_data.jl
@@ -117,7 +117,8 @@ domain =
     Column(; zlim = (FT(-3.0), FT(0.0)), nelements = 10, longlat = (long, lat))
 surface_space = domain.space.surface;
 # Get the paths to each year of MODIS data within the start and stop dates.
-LAI = ClimaLand.prescribed_lai_modis(surface_space, start_date, stop_date);
+LAI =
+    ClimaLand.Canopy.prescribed_lai_modis(surface_space, start_date, stop_date);
 
 # Just like with the air temperature, the LAI is an object that we can use
 # to linearly interpolate observed LAI to any simulation time.
