+++ docs/src/tutorials/integrated/soil_canopy_fluxnet_tutorial.jl
@@ -83,8 +83,11 @@ forcing = FluxnetSimulations.prescribed_forcing_fluxnet(
     FT,
 );
 # LAI for the site - this uses our interface for working with MODIS data.
-LAI =
-    ClimaLand.prescribed_lai_modis(domain.space.surface, start_date, stop_date);
+LAI = ClimaLand.Canopy.prescribed_lai_modis(
+    domain.space.surface,
+    start_date,
+    stop_date,
+);
 
 # # Setup the integrated model
