+++ src/standalone/Vegetation/biomass.jl
@@ -1,0 +1,104 @@
+import ClimaUtilities.TimeVaryingInputs:
+    TimeVaryingInput, LinearInterpolation, PeriodicCalendar
+import Interpolations: Constant
+import ClimaUtilities.Regridders: InterpolationsRegridder
+import ClimaUtilities.FileReaders: NCFileReader, read
+using DocStringExtensions
+
+export prescribed_lai_era5, prescribed_lai_modis
+
+
+
+"""
+     prescribed_lai_era5(era5_lai_ncdata_path,
+                         era5_lai_cover_ncdata_path,
+                         surface_space,
+                         start_date,
+                         earth_param_set;
+                         time_interpolation_method = LinearInterpolation(PeriodicCalendar()),
+                         regridder_type = :InterpolationsRegridder,
+                         interpolation_method = Interpolations.Constant(),)
+
+A helper function which constructs the TimeVaryingInput object for Leaf Area Index, from a
+file path pointing to the ERA5 LAI data in a netcdf file, a file path pointing to the ERA5
+LAI cover data in a netcdf file, the surface_space, the start date, and the earth_param_set.
+
+This currently one works when a single file is passed for both the era5 lai and era5 lai cover data.
+
+The ClimaLand default is to use nearest neighbor interpolation, but
+linear interpolation is supported
+by passing interpolation_method = Interpolations.Linear().
+"""
+function prescribed_lai_era5(
+    era5_lai_ncdata_path,
+    era5_lai_cover_ncdata_path,
+    surface_space,
+    start_date;
+    time_interpolation_method = LinearInterpolation(PeriodicCalendar()),
+    regridder_type = :InterpolationsRegridder,
+    interpolation_method = Interpolations.Constant(),
+)
+    hvc_ds = NCFileReader(era5_lai_cover_ncdata_path, "cvh")
+    lvc_ds = NCFileReader(era5_lai_cover_ncdata_path, "cvl")
+    hv_cover = read(hvc_ds)
+    lv_cover = read(lvc_ds)
+    close(hvc_ds)
+    close(lvc_ds)
+    compose_function = let hv_cover = hv_cover, lv_cover = lv_cover
+        (lai_hv, lai_lv) -> lai_hv .* hv_cover .+ lai_lv .* lv_cover
+    end
+    return TimeVaryingInput(
+        era5_lai_ncdata_path,
+        ["lai_hv", "lai_lv"],
+        surface_space;
+        start_date,
+        regridder_type,
+        regridder_kwargs = (; interpolation_method),
+        method = time_interpolation_method,
+        compose_function = compose_function,
+    )
+end
+
+"""
+     prescribed_lai_modis(surface_space,
+                          start_date
+                          stop_date,
+                          earth_param_set;
+                          time_interpolation_method =
+                                        LinearInterpolation(PeriodicCalendar()))
+                          regridder_type = :InterpolationsRegridder,
+                          interpolation_method = Interpolations.Constant(),
+                          context = ClimaComms.context(surface_space))
+
+A helper function which constructs the TimeVaryingInput object for Leaf Area
+Index using MODIS LAI data; requires the
+surface_space, the start and stop dates, and the earth_param_set.
+
+The ClimaLand default is to use nearest neighbor interpolation, but
+linear interpolation is supported
+by passing interpolation_method = Interpolations.Linear().
+"""
+function prescribed_lai_modis(
+    surface_space,
+    start_date,
+    stop_date;
+    time_interpolation_method = LinearInterpolation(),
+    regridder_type = :InterpolationsRegridder,
+    interpolation_method = Interpolations.Constant(),
+    context = ClimaComms.context(surface_space),
+)
+    modis_lai_ncdata_path = ClimaLand.Artifacts.modis_lai_multiyear_paths(;
+        context,
+        start_date,
+        stop_date,
+    )
+    return TimeVaryingInput(
+        modis_lai_ncdata_path,
+        ["lai"],
+        surface_space;
+        start_date,
+        regridder_type,
+        regridder_kwargs = (; interpolation_method),
+        method = time_interpolation_method,
+    )
+end
