+++ test/integrated/full_land.jl
@@ -206,8 +206,11 @@ forcing = ClimaLand.prescribed_forcing_era5(
     earth_param_set,
     FT;
 )
-LAI =
-    ClimaLand.prescribed_lai_modis(domain.space.surface, start_date, stop_date);
+LAI = ClimaLand.Canopy.prescribed_lai_modis(
+    domain.space.surface,
+    start_date,
+    stop_date,
+);
 
 land = LandModel{FT}(forcing, LAI, toml_dict, domain, Δt);
